floating icons created in Default template and passed down as props
